ai-worker/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pika>=1.3.2
redis>=5.0.1
python-dotenv>=1.0.0
opencv-python>=4.8.0
ultralytics>=8.0.0
numpy>=1.24.0
